backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
requests
python-dotenv